Move affiliations page from User Management to its own submenu
