+++ indico/modules/affiliations/client/js/AffiliationListFilter.tsx
