+++ indico/modules/affiliations/client/js/AffiliationsDashboard.module.scss
