+++ indico/modules/affiliations/client/js/StringListField.tsx
