+++ indico/modules/affiliations/client/js/types.ts
