+++ indico/modules/users/blueprint.py
@@ -8,17 +8,15 @@
 from flask import has_request_context, request
 
 from indico.modules.users.api import RHUserAPI
-from indico.modules.users.controllers import (RHAcceptRegistrationRequest, RHAdmins, RHAffiliationAPI,
-                                              RHAffiliationsAPI, RHAffiliationsDashboard, RHCountries,
-                                              RHExportDashboardICS, RHExportDashboardICSLegacy, RHPersonalData,
-                                              RHPersonalDataUpdate, RHProfilePictureDisplay, RHProfilePicturePage,
-                                              RHProfilePicturePreview, RHRegistrationRequestList,
-                                              RHRejectRegistrationRequest, RHSaveProfilePicture, RHSearchAffiliations,
-                                              RHUserBlock, RHUserDashboard, RHUserDataExport, RHUserDataExportAPI,
-                                              RHUserDataExportDownload, RHUserDelete, RHUserEmails, RHUserEmailsDelete,
-                                              RHUserEmailsSetPrimary, RHUserEmailsVerify, RHUserFavorites,
-                                              RHUserFavoritesAPI, RHUserFavoritesCategoryAPI, RHUserFavoritesEventAPI,
-                                              RHUserPreferences, RHUserPreferencesMarkdownAPI,
+from indico.modules.users.controllers import (RHAcceptRegistrationRequest, RHAdmins, RHExportDashboardICS,
+                                              RHExportDashboardICSLegacy, RHPersonalData, RHPersonalDataUpdate,
+                                              RHProfilePictureDisplay, RHProfilePicturePage, RHProfilePicturePreview,
+                                              RHRegistrationRequestList, RHRejectRegistrationRequest,
+                                              RHSaveProfilePicture, RHUserBlock, RHUserDashboard, RHUserDataExport,
+                                              RHUserDataExportAPI, RHUserDataExportDownload, RHUserDelete, RHUserEmails,
+                                              RHUserEmailsDelete, RHUserEmailsSetPrimary, RHUserEmailsVerify,
+                                              RHUserFavorites, RHUserFavoritesAPI, RHUserFavoritesCategoryAPI,
+                                              RHUserFavoritesEventAPI, RHUserPreferences, RHUserPreferencesMarkdownAPI,
                                               RHUserPreferencesMastodonServer, RHUsersAdmin, RHUsersAdminCreate,
                                               RHUsersAdminMerge, RHUsersAdminMergeCheck, RHUsersAdminSettings,
                                               RHUserSearch, RHUserSearchInfo, RHUserSearchToken,
@@ -93,16 +91,8 @@ _bp.add_url_rule('/search/info', 'user_search_info', RHUserSearchInfo)
 _bp.add_url_rule('/search/', 'user_search', RHUserSearch)
 _bp.add_url_rule('/search/token', 'user_search_token', RHUserSearchToken)
 
-# Affiliations
-_bp.add_url_rule('!/admin/affiliations', 'affiliations_dashboard', RHAffiliationsDashboard)
-_bp.add_url_rule('!/api/admin/affiliations', 'api_admin_affiliations', RHAffiliationsAPI, methods=('GET', 'POST'))
-_bp.add_url_rule('!/api/admin/affiliations/<int:affiliation_id>', 'api_admin_affiliation', RHAffiliationAPI,
-                 methods=('GET', 'PATCH', 'DELETE'))
-_bp.add_url_rule('!/api/countries', 'api_countries', RHCountries)
-
 # Users API
 _bp.add_url_rule('!/api/user/', 'authenticated_user', RHUserAPI)
-_bp.add_url_rule('!/api/affiliations/', 'api_affiliations', RHSearchAffiliations)
 
 
 @_bp.url_defaults
