+++ indico/modules/users/controllers.py
@@ -41,8 +41,7 @@ from indico.modules.events import Event
 from indico.modules.events.contributions.models.contributions import Contribution
 from indico.modules.events.sessions.models.sessions import Session
 from indico.modules.events.util import serialize_event_for_ical
-from indico.modules.logs.models.entries import AppLogRealm, LogKind, UserLogRealm
-from indico.modules.logs.util import make_diff_log
+from indico.modules.logs.models.entries import LogKind, UserLogRealm
 from indico.modules.users import User, logger, user_management_settings
 from indico.modules.users.export_schemas import DataExportRequestSchema
 from indico.modules.users.forms import (AdminAccountRegistrationForm, AdminsForm, AdminUserSettingsForm, MergeForm,
@@ -53,16 +52,15 @@ from indico.modules.users.models.export import DataExportOptions, DataExportRequ
 from indico.modules.users.models.users import ProfilePictureSource, UserTitle
 from indico.modules.users.operations import (add_secondary_email, create_user, delete_or_anonymize_user, grant_admin,
                                              revoke_admin)
-from indico.modules.users.schemas import (AffiliationArgs, AffiliationSchema, BasicCategorySchema, FavoriteEventSchema,
+from indico.modules.users.schemas import (AffiliationSchema, BasicCategorySchema, FavoriteEventSchema,
                                           UserPersonalDataSchema)
 from indico.modules.users.util import (GravatarError, get_avatar_url_from_name, get_gravatar_for_user,
                                        get_linked_events, get_mastodon_server_name, get_related_categories,
                                        get_suggested_categories, get_unlisted_events, get_user_by_email,
-                                       get_user_titles, log_user_update, merge_users, search_affiliations, search_users,
-                                       send_avatar, serialize_user, set_user_avatar)
-from indico.modules.users.views import (WPAffiliationsDashboard, WPUser, WPUserDashboard, WPUserDataExport,
-                                        WPUserFavorites, WPUserPersonalData, WPUserProfilePic, WPUsersAdmin)
-from indico.util.countries import get_countries
+                                       get_user_titles, log_user_update, merge_users, search_users, send_avatar,
+                                       serialize_user, set_user_avatar)
+from indico.modules.users.views import (WPUser, WPUserDashboard, WPUserDataExport, WPUserFavorites, WPUserPersonalData,
+                                        WPUserProfilePic, WPUsersAdmin)
 from indico.util.date_time import now_utc
 from indico.util.i18n import _, force_locale
 from indico.util.images import square
@@ -321,88 +319,6 @@ class RHPersonalDataUpdate(RHUserBase):
             log_user_update(self.user, changes)
         flash(_('Your personal data was successfully updated.'), 'success')
         return '', 204
-
-
-class RHSearchAffiliations(RH):
-    @use_kwargs({'q': fields.String(load_default='')}, location='query')
-    def _process(self, q):
-        res = search_affiliations(q)
-        basic_fields = ('id', 'name', 'code', 'street', 'postcode', 'city', 'country_code', 'meta')
-        return AffiliationSchema(many=True, only=basic_fields).jsonify(res)
-
-
-class RHAffiliationsAPI(RHAdminBase):
-    """List/create affiliations via the admin API."""
-
-    def _process_GET(self):
-        affiliations = (Affiliation.query
-                        .filter(~Affiliation.is_deleted)
-                        .order_by(db.func.indico.indico_unaccent(db.func.lower(Affiliation.name)))
-                        .all())
-        return AffiliationSchema(many=True).jsonify(affiliations)
-
-    @use_args(AffiliationArgs)
-    def _process_POST(self, data):
-        affiliation = Affiliation()
-        affiliation.populate_from_dict(data)
-        db.session.add(affiliation)
-        db.session.flush()
-        signals.affiliations.affiliation_created.send(affiliation)
-        affiliation.log(AppLogRealm.admin, LogKind.positive, 'Affiliation',
-                         f'Affiliation "{affiliation.name}" created', session.user)
-        search_affiliations.bump_version()
-        return AffiliationSchema().jsonify(affiliation), 201
-
-
-class RHAffiliationAPI(RHAdminBase):
-    """CRUD operations on a single affiliation."""
-
-    @use_kwargs({
-        'affiliation': ModelField(Affiliation, filter_deleted=True, required=True, data_key='affiliation_id')
-    }, location='view_args')
-    def _process_args(self, affiliation):
-        RHAdminBase._process_args(self)
-        self.affiliation = affiliation
-
-    def _process_GET(self):
-        return AffiliationSchema().jsonify(self.affiliation)
-
-    @use_args(AffiliationArgs, partial=True)
-    def _process_PATCH(self, data):
-        signals.affiliations.affiliation_updated.send(self.affiliation, payload=data)
-        if not data:
-            return '', 204
-        changes = self.affiliation.populate_from_dict(data)
-        db.session.flush()
-        log_fields = {
-            'name': 'Name',
-            'alt_names': 'Alternative names',
-            'street': 'Street',
-            'postcode': 'Postcode',
-            'city': 'City',
-            'country_code': 'Country',
-            'meta': 'Metadata',
-        }
-        self.affiliation.log(AppLogRealm.admin, LogKind.change, 'Affiliation',
-                                f'Affiliation "{self.affiliation.name}" modified', session.user,
-                                data={'Changes': make_diff_log(changes, log_fields)})
-        search_affiliations.bump_version()
-        return '', 204
-
-    def _process_DELETE(self):
-        self.affiliation.is_deleted = True
-        db.session.flush()
-        self.affiliation.log(AppLogRealm.admin, LogKind.negative, 'Affiliation',
-                             f'Affiliation "{self.affiliation.name}" deleted', session.user)
-        search_affiliations.bump_version()
-        return '', 204
-
-
-class RHCountries(RHAdminBase):
-    """Return the available countries for affiliation forms."""
-
-    def _process(self):
-        return jsonify(list(get_countries().items()))
 
 
 class RHProfilePicturePage(RHUserBase):
@@ -835,13 +751,6 @@ class RHUsersAdmin(RHAdminBase):
                                             num_of_users=num_of_users, num_deleted_users=num_deleted_users,
                                             num_reg_requests=num_reg_requests,
                                             has_moderation=multipass.has_moderated_providers)
-
-
-class RHAffiliationsDashboard(RHAdminBase):
-    """Entry point for the Affiliations admin dashboard."""
-
-    def _process(self):
-        return WPAffiliationsDashboard.render_template('affiliations_dashboard.html', 'affiliations')
 
 
 class RHUsersAdminSettings(RHAdminBase):
