+++ indico/modules/users/module.json
@@ -5,7 +5,6 @@
     "profile_picture": "./ProfilePicture.jsx",
     "personal_data": "./personal_data.js",
     "favorites": "./Favorites.jsx",
-    "data_export": "./DataExport.jsx",
-    "affiliations_dashboard": "./affiliations/index.tsx"
+    "data_export": "./DataExport.jsx"
   }
 }
