+++ indico/modules/users/views.py
@@ -71,8 +71,3 @@ class WPUserDataExport(WPUser):
 class WPUsersAdmin(WPAdmin):
     template_prefix = 'users/'
     bundles = ('module_users.js',)
-
-
-class WPAffiliationsDashboard(WPAdmin):
-    template_prefix = 'users/'
-    bundles = ('module_users.affiliations_dashboard.js', 'module_users.affiliations_dashboard.css')
