+++ indico/web/client/js/react/components/PersonDetailsModal.jsx
@@ -5,7 +5,7 @@
 // modify it under the terms of the MIT License; see the
 // LICENSE file for more details.
 
-import searchAffiliationURL from 'indico-url:users.api_affiliations';
+import searchAffiliationURL from 'indico-url:affiliations.api_affiliations';
 
 import PropTypes from 'prop-types';
 import React, {useState} from 'react';
